Adding a menu for extra calculations
